requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
pymongo==4.6.1
python-dotenv==1.0.0
pandas>=2.2.0
tiktoken>=0.8.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0,<3.0.0
loguru>=0.7.2
redis>=5.0.0
